Update BLAKE3 to 1.5.1 (+ a few)
This updates our BLAKE3 to a few commits past the 1.5.1 release. While
generally it would be ideal to stay on a release, in this case going a
few commits past the release gets us a version of BLAKE3 with an
included Apache 2.0 with LLVM Exceptions license.

This is the commit our Blake is updated to:
https://github.com/BLAKE3-team/BLAKE3/commit/454ee5a7c73583cb3060d1464a5
d3a4e65f06062