[HLSL] Add trait for vector layout compatability
In HLSL data types that are layout compatible with native vectors can
be used in some contexts with resources where other data types are not
allowed because resources read native vectors and convert to the target
type.

To make it easy to express these behaviors it is useful to surface a
type trait.

Related https://github.com/llvm/wg-hlsl/issues/64